remove unused implementations
