+++ src/trace/implementations/mod.rs
@@ -38,13 +38,10 @@
 //! into 'ordered builder'". Then the builder would be bright enough to emit a "batch" for the composite
 //! trace, rather than just a batch of the type merged.
 
-// pub mod spine_fueled;
-pub mod spine_fueled_neu;
-pub use self::spine_fueled_neu as spine_fueled;
+pub mod spine_fueled;
 
 mod merge_batcher;
 
 pub use self::merge_batcher::MergeBatcher as Batcher;
 
 pub mod ord;
-// pub mod hash;
